Allow test_mixed_modulestore to be run without django setup
