+++ common/lib/xmodule/xmodule/modulestore/split_mongo/mongo_connection.py
@@ -13,7 +13,13 @@ from time import time
 
 # Import this just to export it
 from pymongo.errors import DuplicateKeyError  # pylint: disable=unused-import
-from django.core.cache import get_cache, InvalidCacheBackendError
+
+try:
+    from django.core.cache import get_cache, InvalidCacheBackendError
+    DJANGO_AVAILABLE = True
+except ImportError:
+    DJANGO_AVAILABLE = False
+
 import dogstats_wrapper as dog_stats_api
 
 from contracts import check, new_contract
@@ -216,15 +222,16 @@ class CourseStructureCache(object):
     for set and get.
     """
     def __init__(self):
-        self.no_cache_found = False
-        try:
-            self.cache = get_cache('course_structure_cache')
-        except InvalidCacheBackendError:
-            self.no_cache_found = True
+        self.cache = None
+        if DJANGO_AVAILABLE:
+            try:
+                self.cache = get_cache('course_structure_cache')
+            except InvalidCacheBackendError:
+                pass
 
     def get(self, key, course_context=None):
         """Pull the compressed, pickled struct data from cache and deserialize."""
-        if self.no_cache_found:
+        if self.cache is None:
             return None
 
         with TIMER.timer("CourseStructureCache.get", course_context) as tagger:
@@ -245,7 +252,7 @@ class CourseStructureCache(object):
 
     def set(self, key, structure, course_context=None):
         """Given a structure, will pickle, compress, and write to cache."""
-        if self.no_cache_found:
+        if self.cache is None:
             return None
 
         with TIMER.timer("CourseStructureCache.set", course_context) as tagger:
